routes page is present to clean up server
